Cleaned Sidebar, Corrected FlexMasonry and made more progress on the Lightbox
FlexMasonry was behaving incorrectly when the body overflow attribute was changed (mobile sidebar/ lightbox opening). Lightbox now only closes when the "close" tag/'x'/icon is pressed, added a div for potential info about the image and made it so that if you were to click on the left half of the screen it would change to the left image and vice versa. Also added a simple transition between images inside the lightbox.